Move query rewriting to database instrumentation
Fixes #644.
