+++ src/django_mysql/checks.py
@@ -1,5 +1,6 @@
 from django.core.checks import Tags, Warning, register
-from django.db import DEFAULT_DB_ALIAS, connections
+
+from django_mysql.utils import mysql_connections
 
 
 def register_checks():
@@ -70,13 +71,3 @@ def utf8mb4_warning(alias):
         ),
         id="django_mysql.W003",
     )
-
-
-def mysql_connections():
-    conn_names = [DEFAULT_DB_ALIAS] + list(set(connections) - {DEFAULT_DB_ALIAS})
-    for alias in conn_names:
-        connection = connections[alias]
-        if not hasattr(connection, "mysql_version"):
-            continue  # pragma: no cover
-
-        yield alias, connection
