+++ src/django_mysql/utils.py
@@ -299,3 +299,13 @@ def index_name(model, *field_names, **kwargs):
 
 def get_list_sql(sequence):
     return "({})".format(",".join("%s" for x in sequence))
+
+
+def mysql_connections():
+    conn_names = [DEFAULT_DB_ALIAS] + list(set(connections) - {DEFAULT_DB_ALIAS})
+    for alias in conn_names:
+        connection = connections[alias]
+        if connection.vendor != "mysql":
+            continue
+
+        yield alias, connection
